+++ packages/client-connectivity/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [typescript v1.0.0-alpha.2] packages/client-connectivity
 ### Releases
 - [NPM v1.0.0-alpha.2] https://www.npmjs.com/package/@formance/sdk-connectivity/v/1.0.0-alpha.2 - packages/client-connectivity
+
+## 2024-07-11 00:21:58
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.331.0 (2.365.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.3] packages/client-connectivity
+### Releases
+- [NPM v1.0.0-alpha.3] https://www.npmjs.com/package/@formance/sdk-connectivity/v/1.0.0-alpha.3 - packages/client-connectivity